mr. poohohohohohstman

+++ keepr/Services/VaultKeepsService.cs
@@ -13,6 +13,13 @@ namespace keepr.Services
       _repo = repo;
     }
 
+    public VaultKeep Get(int id)
+    {
+      VaultKeep original = _repo.Get(id);
+      if (original == null) { throw new Exception("Invalid Id"); }
+      return original;
+    }
+
     public VaultKeep Create(VaultKeep newVk)
     {
       int id = _repo.Create(newVk);
@@ -20,10 +27,13 @@ namespace keepr.Services
       return newVk;
     }
 
-    internal string Delete(int id)
+    internal string Delete(int id, string userId)
     {
+      VaultKeep original = _repo.Get(id);
+      if (original == null) { throw new Exception("Bad ID"); }
+      if (original.CreatorId != userId) { throw new Exception("Access Denied: Cannot Edit a Vault that you did not create"); }
       _repo.Delete(id);
-      return "Successfully Deleted";
+      return "successfully deleted";
     }
   }
 }
